+++ src/main/java/com/bidnbuy/server/dto/AuctionListResponseDto.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Getter
 @Builder
 public class AuctionListResponseDto {
-    private Integer auctionId;
+    private Long auctionId;
     private Long sellerId;
     private String title;
     private Integer currentPrice; // 현재 최고 입찰가
